fix: 페이지 1부터 처리되도록 변경

+++ community/src/main/java/kakaotech/community/global/page/PageResult.java
@@ -7,6 +7,6 @@ public record PageResult<T>(
     int pageNum,
     int pageSize,
     int totalPage,
-    int totalSize
+    long totalSize
 ) {
 }
